Perf: load dependency before snapshot

+++ snapshot_helper.py
@@ -10,13 +10,14 @@ from common import get_raw_img, get_qcow
 
 
 parser = argparse.ArgumentParser()
+parser.add_argument('--dryrun', default=False, action='store_true')
 parser.add_argument('target', type=str)
 args = parser.parse_args()
 
 setup_work_dir(target=args.target)
 
 # Remove qcow if exists
-if isfile(get_qcow(args.target)):
+if not args.dryrun and isfile(get_qcow(args.target)):
     os.remove(get_qcow(args.target))
 
 # Create qcow
@@ -25,7 +26,11 @@ qemu_img_path = f"{PANDA_BUILD}/qemu-img"
 cmd = [qemu_img_path]
 cmd += ['convert', '-f', 'raw', '-O', 'qcow2',
         get_raw_img(), get_qcow(args.target)]
-subprocess.check_call(cmd)
+    
+if args.dryrun:
+    print(cmd)
+else:
+    subprocess.check_call(cmd)
 
 # Start qemu to obtain snapshot
 cmd = [get_qcow(args.target)]
@@ -33,10 +38,21 @@ cmd += get_extra_args(args.target)
 
 print(" ".join([qemu_path] + cmd))
 
+if args.dryrun:
+    sys.exit()
+
 p = pexpect.spawn(qemu_path, cmd)
 p.logfile = sys.stderr.buffer
+# Login
 p.expect('syzkaller login: ', timeout=600)
 p.sendline('root')
+p.expect('root@syzkaller:~# ')
+# Load dependent kernels
+p.sendline(f'modprobe -v -n {args.target} |sed \$d > load_dep_module.sh')
+p.expect('root@syzkaller:~# ')
+p.sendline('bash load_dep_module.sh')
+p.expect('root@syzkaller:~# ')
+p.sendline('lsmod')
 p.expect('root@syzkaller:~# ')
 # Ctrl-A+C to open qemu console
 p.send('\001c')
